% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{test_down_example}
\alias{test_down_example}
\title{Run a test_down example}
\usage{
test_down_example(
  project_name = NULL,
  author = NULL,
  environment = here::here(),
  pkg = system.file("fakepkg", package = "testdown"),
  book_path = "tests/testdown",
  with_help = TRUE,
  open = interactive()
)
}
\arguments{
\item{project_name}{The name you want to give to the project. The default is `NULL`, which will be then be converted to `basename(here::here())`.}

\item{author}{The author of the test report. Default is set to `NULL`, then it will be skipped.}

\item{environment}{A name for the testing environment. Default is `here::here()`.}

\item{pkg}{The path to the package to document. Default is `here::here()`.}

\item{book_path}{The path to the bookdown output. Default is `"tests/testdown"`.}

\item{with_help}{Should the help appendix be added? Default is `TRUE`.}

\item{open}{Should the bookdown be opened once compiled? Default is `interactive()`.}
}
\value{
Path to the report
}
\description{
Run a test_down example
}
\examples{
if (interactive()){
 dir_book <- tempfile()
 dir.create(dir_book)
 test_down_example(book_path = dir_book)
}
}
