% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomarkdown.R
\name{test_down}
\alias{test_down}
\title{testthat to bookdown}
\usage{
test_down(pkg = ".", book_path = "tests/testdown", open = TRUE)
}
\arguments{
\item{pkg}{The path to the package. Default is `.`}

\item{book_path}{The path to the bookdown output. Default is `"tests/testdown"`.}

\item{open}{Should the bookdown be opened once compiled? Default is TRUE.}
}
\description{
This function turns the results of testthat into a bookdown. Each chapter is a context. The first page gives a summary of all the tests.
}
